create pallet module in primitives
