+++ pallets/storage-provider/src/lib.rs
@@ -57,6 +57,7 @@ pub mod pallet {
     };
     use primitives::{
         commitment::{CommD, CommR, Commitment},
+        pallets::CurrentDeadline,
         proofs::{
             derive_prover_id, Market, ProofVerification, PublicReplicaInfo, Randomness,
             RegisteredPoStProof, SectorNumber, StorageProviderValidation, MAX_SEAL_PROOF_BYTES,
@@ -1075,7 +1076,7 @@ pub mod pallet {
         /// conversion between BlockNumbers fails, but technically should not ever happen.
         pub fn current_deadline(
             storage_provider: &T::AccountId,
-        ) -> Option<primitives::proofs::CurrentDeadline<BlockNumberFor<T>>> {
+        ) -> Option<CurrentDeadline<BlockNumberFor<T>>> {
             let sp = StorageProviders::<T>::try_get(storage_provider).ok()?;
             let current_block = <frame_system::Pallet<T>>::block_number();
 
@@ -1090,7 +1091,7 @@ pub mod pallet {
                 )
                 .ok()?;
 
-            Some(primitives::proofs::CurrentDeadline {
+            Some(CurrentDeadline {
                 deadline_index: deadline.idx,
                 open: deadline.is_open(),
                 challenge_block: deadline.challenge,
