+++ primitives/src/lib.rs
@@ -1,6 +1,7 @@
 #![cfg_attr(not(feature = "std"), no_std)] // no_std by default, requires "std" for std-support
 
 pub mod commitment;
+pub mod pallets;
 pub mod proofs;
 pub mod randomness;
